+++ packages/terra-button-group/tests/jest/ButtonGroupButton.test.jsx
@@ -2,34 +2,42 @@ import React from 'react';
 import Button from '../../src/ButtonGroupButton';
 
 it('should render a default component', () => {
-  const button = shallow(<Button />);
+  const button = shallow(<Button text="Default Button" />);
   expect(button).toMatchSnapshot();
 });
 
-it('should render as selected', () => {
-  const button = shallow(<Button isSelected />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render with text', () => {
-  const button = shallow(<Button text="text" />);
+it('should render as disabled', () => {
+  const button = shallow(<Button text="Disabled" isDisabled />);
   expect(button).toMatchSnapshot();
 });
 
 it('should render with icon only', () => {
   const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} />);
+  const button = shallow(<Button text="Icon Only" icon={testElement} />);
   expect(button).toMatchSnapshot();
 });
 
-it('should render with icon and text', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} text="text" />);
+it('should render with onBlur', () => {
+  const button = shallow(<Button text="onBlur" onBlur={() => {}} />);
   expect(button).toMatchSnapshot();
 });
 
-it('should render with icon and text reversed', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} text="text" isReversed />);
+it('should render with onClick', () => {
+  const button = shallow(<Button text="onClick" onClick={() => {}} />);
+  expect(button).toMatchSnapshot();
+});
+
+it('should render with onFocus', () => {
+  const button = shallow(<Button text="onFocus" onFocus={() => {}} />);
+  expect(button).toMatchSnapshot();
+});
+
+it('should render with onKeyDown', () => {
+  const button = shallow(<Button text="onKeyDown" onKeyDown={() => {}} />);
+  expect(button).toMatchSnapshot();
+});
+
+it('should render with onKeyUp', () => {
+  const button = shallow(<Button text="onKeyUp" onKeyUp={() => {}} />);
   expect(button).toMatchSnapshot();
 });