+++ packages/terra-button-group/tests/nightwatch/ButtonGroupNotSelectable.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import ButtonGroup from '../../lib/ButtonGroup';
+
+export default () => (
+  <ButtonGroup id="button-group-not-selectable" >
+    <ButtonGroup.Button text="Not Selectable 1" key="1" />
+    <ButtonGroup.Button text="Not Selectable 2" key="2" />
+    <ButtonGroup.Button text="Not Selectable 3" key="3" />
+    <ButtonGroup.Button text="Not Selectable 4" key="4" />
+  </ButtonGroup>
+);