+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -3,11 +3,16 @@ import PropTypes from 'prop-types';
 import ReactDatePicker from 'react-datepicker';
 import 'terra-base/lib/baseStyles';
 import ResponsiveElement from 'terra-responsive-element';
+import PopperContainer from './_PopperContainer';
 import DateInput from './DateInput';
 import DateUtil from './DateUtil';
 import styles from './DatePicker.scss';
 
 const propTypes = {
+  /**
+   * Whether the date input should be disabled.
+   */
+  disabled: PropTypes.bool,
   /**
    * An array of ISO 8601 string representation of the dates to disable in the picker.
    */
@@ -72,6 +77,7 @@ const propTypes = {
 };
 
 const defaultProps = {
+  disabled: false,
   excludeDates: undefined,
   filterDate: undefined,
   includeDates: undefined,
@@ -311,12 +317,18 @@ class DatePicker extends React.Component {
         fixedHeight
         locale={intl.locale}
         placeholderText={dateFormat}
+        popperContainer={PopperContainer}
+        popperPlacement="bottom"
+        popperModifiers={{
+          keepTogether: {
+            enabled: true,
+          },
+        }}
         popoverAttachment="top center"
         popoverTargetAttachment="bottom center"
         dropdownMode={'select'}
         showMonthDropdown
         showYearDropdown
-        tetherConstraints={[{ to: 'window', attachment: 'together', pin: true }]}
         name={name}
       />);