+++ packages/terra-dialog/tests/jest/Dialog.test.jsx
@@ -2,18 +2,12 @@ import React from 'react';
 import Dialog from '../../src/Dialog';
 
 describe('Dialog', () => {
-  const defaultRender = <Dialog />;
+  const defaultRender = <Dialog header="Header Content" footer="Footer Content">some body content</Dialog>;
 
-  // Snapshot Tests
+  // Snapshot Test
   it('should render a default component', () => {
     const wrapper = shallow(defaultRender);
     expect(wrapper).toMatchSnapshot();
-  });
-
-  // Prop Tests
-  it('should use the default value when no value is given', () => {
-    const wrapper = shallow(defaultRender);
-    expect(wrapper.find('.dialog').text()).toEqual('default');
   });
 
   // Structure Tests