+++ packages/terra-popup/tests/nightwatch/components/OverlayPopup.jsx
@@ -15,6 +15,9 @@ class DefaultPopup extends React.Component {
     this.overlayStyle = document.documentElement.style.overflow;
     this.overlayId = document.documentElement.id;
     document.documentElement.id = 'popup-overlay-test';
+  }
+
+  componentDidMount() {
     document.documentElement.style.overflow = 'auto';
   }