+++ packages/terra-site/src/examples/embedded-content-consumer/consumers/OnReadyConsumer.jsx
@@ -3,12 +3,8 @@ import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
 
 const OnReadyConsumer = () => (
   <EmbeddedContentConsumer
-<<<<<<< HEAD:packages/terra-site-examples/src/examples/embedded-content-consumer/consumers/OnReadyConsumer.jsx
     src="#/raw/tests/embedded-content-consumer/providers/on-ready-provider"
     fill
-=======
-    src="#/tests/embedded-content-consumer-tests/on-ready-provider"
->>>>>>> master:packages/terra-site/src/examples/embedded-content-consumer/consumers/OnReadyConsumer.jsx
   />
 );