+++ packages/terra-slide-panel/tests/nightwatch/SlidePanelTests.jsx
@@ -16,6 +16,8 @@ const SlidePanelTests = () => (
       <li><Link to="/tests/slide-panel-tests/open">SlidePanel - Open</Link></li>
       <li><Link to="/tests/slide-panel-tests/fill">SlidePanel - Fill</Link></li>
       <li><Link to="/tests/slide-panel-tests/fullscreen">SlidePanel - Fullscreen</Link></li>
+      <li><Link to="/tests/slide-panel-tests/behavior-squish-size-large">SlidePanel - SquishLarge</Link></li>
+      <li><Link to="/tests/slide-panel-tests/toggle">SlidePanel - Toggle</Link></li>
     </ul>
   </div>
 );