+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -4,6 +4,23 @@ ChangeLog
 Unreleased
 -----------------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Uplifted component to use terra-button-group V2
+
+### Removed
+* isSelectable prop in CollapsibleMenuViewItemGroup.
+
+### Added
+* selectedKeys prop in CollapsibleMenuViewItemGroup.
+
+1.10.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.9.0 - (January 18, 2018)
 ------------------
 ### Changed