+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,20 @@ ChangeLog
 Unreleased
 ----------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Uplifted component to use terra-button v2
+
+### Added
+* Added `disabled` prop
+
+1.7.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.6.0 - (January 18, 2018)
 ------------------
 ### Changed