+++ packages/terra-dialog/CHANGELOG.md
@@ -4,3 +4,5 @@ ChangeLog
 Unreleased
 -----------------
 * Initial stable release
+* New component, examples, doc, tests.
+* Added theme variables.