+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -1,8 +1,20 @@
 ChangeLog
 =========
 
-(Unreleased)
+Unreleased
 ----------
+
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Update theming values to initial browser default for various properties
+* Merged border-color, border-width, and border-style into short-hand syntax
+
+1.2.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
 
 1.1.0 - (January 18, 2018)
 ------------------