+++ packages/terra-form-radio/CHANGELOG.md
@@ -4,6 +4,18 @@ ChangeLog
 Unreleased
 ------------------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Update theming values to initial browser default for various properties
+* Merged border-color, border-width, and border-style into short-hand syntax
+
+1.1.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.0.0 - (January 18, 2018)
 -----------------
 * Initial stable release