+++ packages/terra-image/CHANGELOG.md
@@ -3,7 +3,13 @@ Changelog
 
 Unreleased
 ----------
-1.15.1
+
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.16.0 - (February 1, 2018)
 ------------------
 ### Changed
 * Add code usage example to README.md