+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 Unreleased
 ----------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Changed React.Children.map to React.Children.forEach to prevent '.$' being prepended to the key.
+
+1.13.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.12.0 - (January 18, 2018)
 ------------------
 ### Added