+++ packages/terra-props-table/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.11.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.10.0 - (January 18, 2018)
 ------------------
 ### Changed