+++ packages/terra-search-field/CHANGELOG.md
@@ -4,6 +4,20 @@ Changelog
 Unreleased
 ----------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Uplifted component to use terra-button v2
+
+### Fixed
+* Fixed error when handle search with a null searchText.
+
+1.20.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.19.0 - (January 18, 2018)
 ------------------
 ### Changed