+++ packages/terra-table/CHANGELOG.md
@@ -3,8 +3,16 @@ Changelog
 
 Unreleased
 ----------
-###Changed
-* Moved table sort indicator 
+
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.20.0 - (February 1, 2018)
+------------------
+### Changed
+* Moved table sort indicator
 
 1.19.0 - (January 18, 2018)
 ------------------