+++ packages/terra-toggle-button/CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 Unreleased
 ----------
 
+2.0.0 - (February 12, 2018)
+------------------
+### Changed
+* Updated to use React 16
+* Uplifted component to use terra-button v2
+
+1.19.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.18.0 - (January 18, 2018)
 ------------------
 ### Changed